move `remove_redundant_poi` to grisubal::model module
