+++ honeycomb-kernels/src/grisubal/kernel.rs
@@ -12,7 +12,7 @@ use std::{
     process::id,
 };
 
-use crate::{Geometry2, GeometryVertex, GridCellId, IsBoundary, MapEdge};
+use crate::{remove_redundant_poi, Geometry2, GeometryVertex, GridCellId, IsBoundary, MapEdge};
 use honeycomb_core::{
     CMap2, CMapBuilder, CoordsFloat, DartIdentifier, EdgeIdentifier, Vertex2, VertexIdentifier,
     NULL_DART_ID,
@@ -131,18 +131,6 @@ pub fn build_mesh<T: CoordsFloat>(
 }
 
 // --- main kernels steps
-
-fn remove_redundant_poi<T: CoordsFloat>(geometry: &mut Geometry2<T>, (cx, cy): (T, T)) {
-    // PoI that land on the grid create a number of issues; removing them is ok since we're intersecting the grid
-    // at their coordinates, so the shape will be captured via intersection anyway
-    geometry.poi.retain(|idx| {
-        let v = geometry.vertices[*idx];
-        // origin is assumed to be (0.0, 0.0)
-        let on_x_axis = (v.x() % cx).is_zero();
-        let on_y_axis = (v.y() % cy).is_zero();
-        !(on_x_axis | on_y_axis)
-    });
-}
 
 #[allow(
     clippy::too_many_lines,
