+++ honeycomb-kernels/src/grisubal/model.rs
@@ -201,6 +201,18 @@ impl<T: CoordsFloat> From<Vtk> for Geometry2<T> {
     }
 }
 
+pub fn remove_redundant_poi<T: CoordsFloat>(geometry: &mut Geometry2<T>, (cx, cy): (T, T)) {
+    // PoI that land on the grid create a number of issues; removing them is ok since we're intersecting the grid
+    // at their coordinates, so the shape will be captured via intersection anyway
+    geometry.poi.retain(|idx| {
+        let v = geometry.vertices[*idx];
+        // origin is assumed to be (0.0, 0.0)
+        let on_x_axis = (v.x() % cx).is_zero();
+        let on_y_axis = (v.y() % cy).is_zero();
+        !(on_x_axis | on_y_axis)
+    });
+}
+
 #[derive(Clone, Debug, PartialEq, Eq, Hash)]
 pub enum GeometryVertex {
     /// Regular vertex. Inner `usize` indicates the vertex ID in-geometry.
