+++ honeycomb-kernels/src/lib.rs
@@ -35,4 +35,6 @@ pub use grisubal::{grisubal, model::Clip};
 #[allow(unused)]
 pub(crate) use grisubal::grid::{BBox2, GridCellId};
 #[allow(unused)]
-pub(crate) use grisubal::model::{Geometry2, GeometryVertex, IsBoundary, MapEdge};
+pub(crate) use grisubal::model::{
+    remove_redundant_poi, Geometry2, GeometryVertex, IsBoundary, MapEdge,
+};
